+++ test/ecto/adapters/clickhouse/connection_test.exs
@@ -984,19 +984,19 @@ defmodule Ecto.Adapters.ClickHouse.ConnectionTest do
 
   test "json_extract_path" do
     query = Schema |> select([s], json_extract_path(s.meta, [0, 1]))
-    assert all(query) == ~s{SELECT JSON_QUERY(s0."meta", '$[0][1]') FROM "schema" AS s0}
+    assert all(query) == ~s{SELECT s0."meta"[0][1] FROM "schema" AS s0}
 
     query = Schema |> select([s], json_extract_path(s.meta, ["a", "b"]))
-    assert all(query) == ~s{SELECT JSON_QUERY(s0."meta", '$.a.b') FROM "schema" AS s0}
+    assert all(query) == ~s{SELECT s0."meta".a.b FROM "schema" AS s0}
 
     query = Schema |> select([s], json_extract_path(s.meta, ["'a"]))
-    assert all(query) == ~s{SELECT JSON_QUERY(s0."meta", '$.''a') FROM "schema" AS s0}
+    assert all(query) == ~s{SELECT s0."meta".`'a` FROM "schema" AS s0}
 
     query = Schema |> select([s], json_extract_path(s.meta, ["\"a"]))
-    assert all(query) == ~s{SELECT JSON_QUERY(s0."meta", '$.\\"a') FROM "schema" AS s0}
+    assert all(query) == ~s{SELECT s0."meta".`"a` FROM "schema" AS s0}
 
     query = Schema |> select([s], s.meta["author"]["name"])
-    assert all(query) == ~s{SELECT JSON_QUERY(s0."meta", '$.author.name') FROM "schema" AS s0}
+    assert all(query) == ~s{SELECT s0."meta".author.name FROM "schema" AS s0}
   end
 
   test "nested expressions" do